Create data_manipulation_queries.sql
Added queries for Draft 4.

After doing the Edit/Update function for Jobs, I felt it was easier to include a yes & no option for is_active (I will change this in the UI). Right now, the one checkbox does not allow for is_active to revert back to no.

For Search Jobs, the only part I am confused about whether to search companies by name or id. Is it easiest if the user enters a company_id rather than name because company_id is an existing column in Jobs? That's how I did the query above. In that case, we just need to display the company_id in the tables up top so the user know what id corresponds to which company. Otherwise I think we will need to find a way to take a company name user input and convert it to an id behind the scenes in Flask